% Generated by roxygen2 (4.0.2): do not edit by hand
\name{agreguj_pwe}
\alias{agreguj_pwe}
\title{Oblicza PWE na podstawie PV agregujac wg wskazanych zmiennych}
\usage{
agreguj_pwe(dane, kolGrupy, bladZrwn, wariancjaPop,
  kolPV = "^[pP][vV]_[0-9]+|wynik$", kolNrPV = "nr_pv")
}
\arguments{
\item{dane}{macierz z wartosciami PV do zagregowania}

\item{kolGrupy}{nazwa kolumny (lub wektor nazw kolumn), wg których PWE mają zostać zagregowane}

\item{bladZrwn}{błąd zrównywania}

\item{wariancjaPop}{wariancja wyników w populacji (uzyskana np. z funkcji pobierz_wariancje_populacji())}

\item{kolPV}{wyrażenie regularne dopasowujące kolumny z wartościami PV}

\item{kolNrPV}{nazwa kolumny z numerem PV (gdy dane w postaci długiej)}
}
\value{
[data.frame] zagregowane wartości PWE
}
\description{
_
}

