% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czy_postac_dluga.R
\name{czy_postac_dluga}
\alias{czy_postac_dluga}
\title{Sprawdza, czy wyniki testow/egzaminow sa w postaci dlugiej}
\usage{
czy_postac_dluga(dane)
}
\arguments{
\item{dane}{wynik dzialania dowolnej z funkcji pobierz_wyniki_...()}
}
\description{
Postac dluga musi zawierac kolumny \code{kryterium} oraz \code{ocena}
}
