% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czy_postac_szeroka.R
\name{czy_postac_szeroka}
\alias{czy_postac_szeroka}
\title{Sprawdza, czy wyniki testow/egzaminow sa w postaci szerokiej}
\usage{
czy_postac_szeroka(dane)
}
\arguments{
\item{dane}{wynik dzialania dowolnej z funkcji pobierz_wyniki_...()}
}
\description{
Postac szeroka musi zawierac kolumny z prefiksem \code{k_} lub \code{p_}
}

