% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_pwe.R
\name{.oblicz_statystyki_pwe}
\alias{.oblicz_statystyki_pwe}
\title{Oblicza statystyki PWE zapisywane w bazie danych dla pojedynczego wektora PV i pojedynczej grupy}
\usage{
.oblicz_statystyki_pwe(x)
}
\arguments{
\item{x}{wektor pojedynczych oszacowań PV dla oblicznaego agregatu}
}
\value{
[data.frame] obliczone statystyki
}
\description{
Oblicza statystyki PWE zapisywane w bazie danych dla pojedynczego wektora PV i pojedynczej grupy
}
