% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_pwe.R
\name{.oblicz_z_pv}
\alias{.oblicz_z_pv}
\title{Oblicza statystyki PWE zapisywane w bazie uśredniając po numerze PV}
\usage{
.oblicz_z_pv(d, blZrwn2, varPop, ilePV)
}
\arguments{
\item{d}{ramka danych z wartości oszacowań PV oraz numerami PV}

\item{blZrwn2}{kwadrat błędu zrównywania (jeśli niewliczony w PV)}

\item{varPop}{wariancja w populacji}

\item{ilePV}{liczba PV użytych do obliczeń}
}
\value{
[data.frame] obliczone statystyki
}
\description{
Oblicza statystyki PWE zapisywane w bazie uśredniając po numerze PV
}
