% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawdz_dokumentacje_zmiennych.R
\name{.sprawdz_dokumentacje_zmiennych}
\alias{.sprawdz_dokumentacje_zmiennych}
\title{Zbiera nazwy wszystkich zmiennych zwracanych przez funkcje pobierz_...()}
\usage{
.sprawdz_dokumentacje_zmiennych(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Funkcja używana do utrzymywania aktualności dokumentacji grup danych oraz
weryfikacji, czy wszystkie funkcje pobierające (wyłączywszy te pobierające
wyniki w postaci szerokiej) odpalają się w domyślnej parametryzacji.
}
