% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{e}
\alias{e}
\title{Konwertuje łańcuch znaków na lokalne kodowanie komputera}
\usage{
e(s)
}
\arguments{
\item{s}{wektor lub macierz do konwersji}
}
\description{
Konwertuje łańcuch znaków na lokalne kodowanie komputera
}

