% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_exam_results}
\alias{get_exam_results}
\title{Pobiera wyniki części egzaminu}
\usage{
get_exam_results(src, rodzajEgzaminu, czescEgzaminu, rokEgzaminu, czyEwd,
  punktuj = TRUE, idSkali = NULL, skroc = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{rodzajEgzaminu}{rodzaj egzaminu, ktorego wyniki maja zostac pobrane}

\item{czescEgzaminu}{czesc egzaminu, ktorego wyniki maja zostac pobrane}

\item{rokEgzaminu}{rok egzaminu, ktorego wyniki maja zostac pobrane}

\item{czyEwd}{wybor, czy maja byc pobrane wyniki gromadzone przez EWD, czy PAOU}

\item{punktuj}{wybor, czy dane maja byc pobrane w postaci dystraktorow, czy punktow}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}
}
\description{
Pobiera wyniki części egzaminu
}

