% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_pupils_estimates}
\alias{get_pupils_estimates}
\title{Pobiera oszacowania umiejętności uczniów}
\usage{
get_pupils_estimates(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera oszacowania umiejętności uczniów
}

