% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_pupils_tests_data}
\alias{get_pupils_tests_data}
\title{Pobiera informacje o uczniach specyficzne dla testu / egzaminu}
\usage{
get_pupils_tests_data(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera informacje o uczniach specyficzne dla testu / egzaminu
}

