% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_test_results}
\alias{get_test_results}
\title{Pobiera wyniki testu}
\usage{
get_test_results(src, idTestu, punktuj = TRUE, idSkali = NULL,
  skroc = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{idTestu}{identyfikator testu do pobrania}

\item{punktuj}{wybor, czy dane maja byc pobrane w postaci dystraktorow, czy punktow}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}
}
\description{
Pobiera wyniki testu
}

