% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizuj.R
\name{normalizuj}
\alias{normalizuj}
\title{Normalizuje wyniki}
\usage{
normalizuj(dane, src = NULL, kolWynik = "wynik", idSkali = NULL,
  skalowanie = NULL, grupa = NULL, ...)
}
\arguments{
\item{dane}{ramka danych zawierająca wyniki}

\item{src}{uchwyt źródła danych dplyr-a (gdy normalizacja na podstawie norm w
bazie)}

\item{kolWynik}{nazwa kolumny zawierającej wyniki}

\item{idSkali}{skala, której norma z bazy ma zostać zastosowana}

\item{skalowanie}{skalowanie w ramach wybranej skali, którego norma z bazy ma
zostać zastosowana}

\item{grupa}{grupa w ramach wybranego skalowania, którego norma z bazy ma
zostać zastosowana}

\item{...}{ew. parametry funkcji normy_ekwikwantylowe() (gdy normalizacja na
podstawie danych)}
}
\description{
Znormalizowane wyniki doklejane są do danych jako nowa zmienna z sufiksem
"_norm".

Wyniki normalizowane są dna jeden z dwóch sposobów - albo ekwikwantylowo na 
podstawie przekazanych danych (gdy nie podano parametru \code{src}) albo za
pomoca wskazanej normy w bazie danych (podajac parametry \code{src},
\code{idSkali}, \code{skalowanie} i ew. \code{grupa}).
}
\seealso{
normy_ekwikwantylowe
}

