% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_informacje_kryterium.R
\name{oblicz_informacje_kryterium}
\alias{oblicz_informacje_kryterium}
\title{Oblicza wartosci funkcji informacyjnej (pseudo)kryterium oceny}
\usage{
oblicz_informacje_kryterium(parametry, theta)
}
\arguments{
\item{parametry}{ramka danych z parametrami (pseudo)kryterium oceny (np.
pobrana za pomocą pobierz_parametry())}

\item{theta}{wektor wartości poziomu umiejętności uczniów, dla którego
obliczona zostanie funkcja informacji}
}
\description{
Ramka danych \code{parametry} musi zawierać kolumny \code{kryterium,
parametr, model, wartosc}.

Zwraca ramkę danych z kolumnami \code{theta} oraz \code{informacja}.
}
