% Generated by roxygen2 (4.0.1): do not edit by hand
\name{oblicz_wariancje_populacji}
\alias{oblicz_wariancje_populacji}
\title{Oblicza wariancję wyników w populacji na podstawie PV}
\usage{
oblicz_wariancje_populacji(dane, kolPV = "^[pP][vV]_[0-9]+|wynik$",
  kolNrPV = "nr_pv")
}
\arguments{
\item{dane}{macierz z wartosciami PV}

\item{kolPV}{wyrażenie regularne dopasowujące kolumny z wartościami PV}

\item{kolNrPV}{nazwa kolumny z numerem PV (gdy dane w postaci długiej)}
}
\value{
[numeric] obliczona wariancja
}
\description{
_
}

