% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/odkoduj_dystraktory.R
\name{odkoduj_dystraktory}
\alias{odkoduj_dystraktory}
\title{Zamienia liczbowe kody dystraktorów na kody literowe}
\usage{
odkoduj_dystraktory(src, dane, kolDystr = "^(odpowiedz|k_[0-9]+)$",
  kolKryt = "^kryterium$")
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{dane}{ramka danych z wynikami uczniów}

\item{kolDystr}{wyrażenie regularne dopasowujące nazwy kolumn z kodami
liczbowymi dystraktorów do zamiany na kody literowe}

\item{kolKryt}{wyrażenie regularne dopasowujęce nazwę kolumny z
identyfikatorem kryterium oceny (tylko dane w postaci długiej)}
}
\description{
Z powodów wydajnościowych dystraktory w bazie zapisywane są w postaci kodów
liczbowych i tak też pobierają je funkcje \code{pobierz_wyniki_...()}.
Funkcja \code{odkoduj_dystraktory()} zamienia kody liczbowe na oznaczenia
literowe użyte pierwotnie w arkuszach testowych (złączając przekazane wyniki
z danymi zwracanymi przez funkcję \code{pobierz_schemary_odp()}).

Z uwagi na brak przenośności niektórych operacji pomiędzy kodem R i SQL, jak
również z uwagi na to, że operacja zamiany po stronie bazy danych na danych
w postaci szerokiej byłaby bardzo kosztowna obliczeniowo, funkcja pobiera
najpierw przekazane dane za pomocą \code{collect()}

Kody liczbowe -1 i -2 oznaczają, odpowiednio, opuszczenie i wielokrotne
zaznaczenie.
}