% Generated by roxygen2 (4.0.2): do not edit by hand
\name{oszacuj_czas_wykonania}
\alias{oszacuj_czas_wykonania}
\title{Pobiera dozwolone wartości punktowe kryteriów oceny}
\usage{
oszacuj_czas_wykonania(dane, pelnyPlan = FALSE, format = "TEXT")
}
\arguments{
\item{dane}{ramka danych dplyr-a}

\item{pelnyPlan}{czy zwracać pełny plan zapytania czy tylko łączny koszt (zawsze TRUE gdy format inny niż TEXT)}

\item{format}{format zwracanych danych: TEXT, XML, JSON lub YAML}
}
\description{
Zwraca albo pełen plan zapytania (pelnyPlan = TRUE) albo skrótowe
podsuomowanie szacowanego kosztu wykonania zapytania.

W wypadku zwracania tylko podsumowania wyswietlane są dolny i górny szacunek
w jednostkach szacowania czasu planera wraz z komentarzem opisującym rząd
wielkości czasu, jakiemu dana wartość odpowiada.
}

