\name{pobierz_dane_jednoroczne_gimnazjum}
\alias{pobierz_dane_jednoroczne_gimnazjum}
\title{Pobiera ramke danych ze scalonymi danymi jednorocznymi egzaminu gimnazjalnego}
\usage{
  pobierz_dane_jednoroczne_gimnazjum(rok,
    zrodloDanychODBC = "EWD")
}
\arguments{
  \item{rok}{rok egzaminu, dla ktorego maja zostac pobrane
  dane}

  \item{zrodloDanychODBC}{nazwa zrodla danych ODBC, ktorego
  nalezy uzyc}
}
\value{
  data frame
}
\description{
  aaa
}
\details{
  bbb
}
\examples{
a=1
}

