% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_dane_uczniowie_testy}
\alias{pobierz_dane_uczniowie_testy}
\title{Pobiera informacje o uczniach specyficzne dla testu / egzaminu}
\usage{
pobierz_dane_uczniowie_testy(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera informacje o uczniach specyficzne dla testu / egzaminu
}

