% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_kryteria_oceny.R
\name{pobierz_kryteria_oceny}
\alias{pobierz_kryteria_oceny}
\title{Pobiera informacje o kryteriach oceny i pytaniach oraz w których
  testach występują}
\usage{
pobierz_kryteria_oceny(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Każde kryterium oceny występuje w danych tyle razy, w ilu
różnych testach występuje.
}

