% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_odpowiedzi}
\alias{get_answers}
\alias{pobierz_odpowiedzi}
\title{Pobiera wyniki testow (odpowiedzi) w postaci dlugiej}
\usage{
pobierz_odpowiedzi(src, idSkali = NULL, skroc = FALSE)

get_answers(src, idSkali = NULL, skroc = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}
}
\description{
Poniewaz skale definiuja przeksztalcenia wynikow w punktach (i nie da sie
tych przeksztalcen zastosowac dla dystraktorow), zastosowanie skali powoduje
ze zwrocony zbior danych pozbawiony bedzie kolumny "odpowiedz"

Zastosowanie skali spowoduje rowniez automatyczne przyciecie kryteriow
jedynie do tych, ktore naleza do wskazanej skali.
}

