% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_parametry}
\alias{pobierz_parametry}
\title{Pobiera parametry kryteriów i pseudokryteriów obliczone w skalowaniach}
\usage{
pobierz_parametry(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera parametry kryteriów i pseudokryteriów obliczone w skalowaniach
}

