% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pobierz_pseudokryteria}
\alias{pobierz_pseudokryteria}
\title{Zwraca wektor pozwalajacy identyfikowac pseudokryteria}
\usage{
pobierz_pseudokryteria(P)
}
\arguments{
\item{P}{otwarte połączenie ODBC}
}
\value{
[character] wektor pseudokryteriow znajdujacych sie w bazie
}
\description{
Nazwy elementow zwracanego wektora to id_pseudokryterium
Elementy zwracanego wektora to posortowane rosnaco id_kryterium
nalezacych do danego pseudokryterium, zlaczone w lancuch znakow
separatorem '|'
}
\details{
_
}

