% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_schematy_pkt}
\alias{pobierz_schematy_pkt}
\title{Pobiera dozwolone wartości punktowe kryteriów oceny}
\usage{
pobierz_schematy_pkt(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera dozwolone wartości punktowe kryteriów oceny
}

