% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_skalowania_elementy_kowariancje.R
\name{pobierz_skalowania_elementy_kowariancje}
\alias{pobierz_skalowania_elementy_kowariancje}
\title{Pobiera dane z tablicy skalowania_elementy_kowariancje}
\usage{
pobierz_skalowania_elementy_kowariancje(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera dane z tablicy skalowania_elementy_kowariancje
}
