% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_szkoly}
\alias{get_schools}
\alias{pobierz_szkoly}
\title{Pobiera informacje o szkołach}
\usage{
pobierz_szkoly(src)

get_schools(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera informacje o szkołach
}

