% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_test}
\alias{pobierz_test}
\title{Pobiera ramke danych z wynikami egzaminacyjnymi wskazanego testu}
\usage{
pobierz_test(idTestu, punktuj = TRUE, idSkali = NULL, skroc = TRUE,
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{idTestu}{identyfikator, ktorego wyniki maja zostac pobrane}

\item{punktuj}{wybor, czy dane maja byc pobrane w postaci dystraktorow, czy punktow}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC, ktorego nalezy uzyc}
}
\value{
data frame
}
\description{
_
}

