% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_testy}
\alias{get_tests}
\alias{pobierz_testy}
\title{Pobiera informacje o testach / egzaminach}
\usage{
pobierz_testy(src)

get_tests(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera informacje o testach / egzaminach
}

