% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_uczniow.R
\name{pobierz_uczniow}
\alias{pobierz_uczniow}
\title{Pobiera informacje o uczniach}
\usage{
pobierz_uczniow(src, daneOsobowe = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{daneOsobowe}{czy pobierac z bazy takze dane osobowe (wymaga specjalnych uprawnien)}
}
\description{
Pobiera informacje o uczniach
}

