% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wartosci_wskaznikow.R
\name{pobierz_wartosci_wskaznikow}
\alias{pobierz_wartosci_wskaznikow}
\title{Ppobiera wartości wskaźników (EWD i PWE)}
\usage{
pobierz_wartosci_wskaznikow(src, czyPomin = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{czyPomin}{czy uwzględniać wartości wskaźników oznaczone jako do pominięcia}
}
\description{
Ppobiera wartości wskaźników (EWD i PWE)
}
