% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_wartosci_wskaznikow_lu}
\alias{pobierz_wartosci_wskaznikow_lu}
\title{Pobiera informacje o liczbie uczniów uwzględnionych przy wyliczaniu
  wskaźników w podziale na przedmioty (dokładniej części egzaminu)}
\usage{
pobierz_wartosci_wskaznikow_lu(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera informacje o liczbie uczniów uwzględnionych przy wyliczaniu
  wskaźników w podziale na przedmioty (dokładniej części egzaminu)
}

