% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wskazniki.R
\name{pobierz_wskazniki}
\alias{pobierz_wskazniki}
\title{Zwraca wskaźniki zapisane w bazie}
\usage{
pobierz_wskazniki(src, doPrezentacji = TRUE, wszystkieKolumny = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{doPrezentacji}{czy pobierać tylko wskaźniki oznaczone do publicznej
prezentacji (TRUE - pobierz tylko publiczne, FALSE - tylko niepuliczne, NA
- wszystkie)}

\item{wszystkieKolumny}{czy pobierać także techniczne kolumny używane do
prezentacji wskaźników na stronach WWW}
}
\description{
Każdy wskaźnik jest powielony w zwracanej tabeli tyle razy, ile wynosi 
iloczyn liczby typów szkół, do których ma zastosowanie, liczby częśći 
egzaminu, do których ma zastosowanie, liczby okresów, dla których wskaźnik
był obliczany oraz liczby skalowań, na których wynikach się opiera. Dzięki
temu możliwe jest łatwe filtrowanie po typach szkół i/lub przedmiotach bez
konieczności tworzenia dodatkowych grup danych. Aby uzyskać listę, w której
kolumna "id_wskaznika" ma unikalne wartości nelaży usunąć kolumny
"typ_szkoly" i "przedmiot" oraz skorzystać z czasownika "distinct()".

Domyśle odfiltrowywanie jedynie wskaźników do prezentacji służy ułatwieniu 
wyszukiwania wskaźników przez większość użytkowników.

Parametr "wszystkieKolumny" pozwala pobrać również kolumny mające jedynie 
zastosowania techiczne, np. kolor prezntacji na stronie WWW, dopełniacz nazwy
egzaminu wyświetlany przy osi wykresu na stronie WWW, itp.
}
