% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wyniki_egzaminu.R
\name{pobierz_wyniki_egzaminu}
\alias{pobierz_wyniki_egzaminu}
\title{Pobiera wyniki części egzaminu}
\usage{
pobierz_wyniki_egzaminu(src, rodzajEgzaminu, czescEgzaminu, rokEgzaminu, czyEwd,
  punktuj = TRUE, idSkali = NULL, skroc = TRUE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{rodzajEgzaminu}{rodzaj egzaminu, ktorego wyniki maja zostac pobrane}

\item{czescEgzaminu}{czesc egzaminu, ktorego wyniki maja zostac pobrane}

\item{rokEgzaminu}{rok egzaminu, ktorego wyniki maja zostac pobrane}

\item{czyEwd}{wybor, czy maja byc pobrane wyniki gromadzone przez EWD, czy PAOU}

\item{punktuj}{wybor, czy dane maja byc pobrane w postaci dystraktorow, czy punktow}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}
}

