% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wyniki_testu.R
\name{pobierz_wyniki_testu}
\alias{pobierz_wyniki_testu}
\title{Pobiera wyniki testu}
\usage{
pobierz_wyniki_testu(
  src,
  idTestu,
  punktuj = TRUE,
  idSkali = NULL,
  skroc = TRUE
)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{idTestu}{identyfikator testu do pobrania}

\item{punktuj}{wybor, czy dane maja byc pobrane w postaci dystraktorow, czy punktow}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}
}
\description{
Pobiera wyniki testu w postaci szerokiej
}
