% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polacz.R
\name{polacz}
\alias{polacz}
\title{Zwraca połączenie dplyr-a z bazą danych ZPD}
\usage{
polacz(
  user = NULL,
  password = NULL,
  dbname = "ewd",
  host = "194.54.27.158",
  port = 5432,
  ...
)
}
\arguments{
\item{user}{użytkownik}

\item{password}{hasło}

\item{dbname}{nazwa bazy danych}

\item{host}{adres serwera bazy danych}

\item{port}{port, na którym nasłuchuje serwer baz danych}

\item{...}{dodatkowe parametry do przekazania do DBI::dbConnect()}
}
\description{
Zwraca połączenie dplyr-a z bazą danych ZPD
}
