% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/przygotuj_mature_popr.R
\name{przygotuj_mature_popr}
\alias{przygotuj_mature_popr}
\title{Przygotowuje zbiory matur poprawkowych do importu do bazy}
\usage{
przygotuj_mature_popr(rok, przedmiot, plik, kodowanie = "Windows-1250",
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{rok}{rok egzaminu}

\item{przedmiot}{przedmiot egzaminacyjny}

\item{plik}{plik CSV z wynikami matury poprawkowej}

\item{kodowanie}{kodowanie używane przy otwieraniu pliku}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC}
}
\value{
void
}
\description{
Przygotowuje zbiory z wynikami matur poprawkowych do importu do bazy
skryptem 'import_test.php'.
Sprawdza zgodność kolumn w pliku z planem testu w bazie, jeśli trzeba,
zamienia kody dystraktorów (angielski i problem T/F v. A/B), a następnie
zmienia ich nazwy na stosowne 'k_idKryterium'.
}

