% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ramka_danych_pv}
\alias{ramka_danych_pv}
\title{Upewnia, że ramka danych z wartościami PV jest w postaci długiej.}
\usage{
ramka_danych_pv(dane, kolPV = "^[pP][vV]_[0-9]+|wynik$", kolNrPV = "nr_pv",
  kolPozostale = NULL)
}
\arguments{
\item{dane}{ramka danych z wartosciami PV}

\item{kolPV}{wyrażenie regularne dopasowujące kolumny z wartościami PV}

\item{kolNrPV}{nazwa kolumny z numerem PV (gdy dane w postaci długiej)}

\item{kolPozostale}{nazwy wszystkich innych kolumn, które mają zostać zachowane}
}
\value{
[data.frame] dane w postaci długiej (z kolumnami "nr_pv" i "wynik")
}
\description{
_
}

