% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sprawdz_podwojne_ocenianie.R
\name{sprawdz_podwojne_ocenianie}
\alias{sprawdz_podwojne_ocenianie}
\title{Sprawdza poprawność pliku wyników podwójnego oceniania}
\usage{
sprawdz_podwojne_ocenianie(plik, codebook, dane, zeszyt = NULL,
  wzor = "^z[0-9].*_ost$")
}
\arguments{
\item{plik}{ścieżka do pliku z danymi podwójnego oceniania (plik powinien mieć "format IBE")}

\item{codebook}{ścieżka do pliku codebooka danych z podwójnego oceniania (plik powinien mieć "format IBE")}

\item{dane}{ramka danych z wynikami uczniów}

\item{zeszyt}{umożliwia odfiltrowanie tylko wskazanych zeszytów - potrzebny do weryfikacji danych gimnazjalnych}

\item{wzor}{wyrażenie regularne dopasowujący nazwy zmiennych z ostatecznymi wartościami ocen}
}
\value{
[logical] TRUE, gdy dane zweryfikowane poprawnie, FALSE w p.p.
}
\description{
Sprawdzane są:
\itemize{
	\item zgodność w ramach pliku podwójnego oceniania, tzn. czy w wypadku dwóch
  zgodnych ocen egzaminatorów ocena ostateczna też jest im równa
	\item zgodność ostatecznej oceny z pliku podwójnego oceniania z danymi w
  zbiorze uczniów
	\item zawieranie się wszystkich identyfikatorów uczniów ze zbioru podwójnego
  oceniania w zbiorze identyfikatorów uczniów ze zbioru uczniów
}

Dane dopasowywane są na podstawie kolumny "id_ucz".

Odpowiadające sobie kolumny z ocenami powinny kończyć się sufiksami:
\itemize{
	\item "_ost" - ocena ostateczna w zbiorze podwójnego oceniania
	\item "_i" - ocena pierwszego egzaminatora w zbiorze podwójnego oceniania
	\item "_ii" - ocena drugiego egzaminatora w zbiorze podwójnego oceniania
	\item bez sufiksu - ocena w zbiorze uczniowskim
}
}
\examples{
\dontrun{
	# zwykłe sprawdzenie
	dane = wczytaj_ibe('spr.csv', 'spr_c.csv')
	sprawdz_podwojne_ocenianie('spr_podw_kod.csv', 'spr_podw_kod_c.csv', dane)

	# dane gimnazjalne - oddzielnie dla części hum i mat-przyr
	dane = wczytaj_ibe('gim.csv', 'gim_c.csv')
	sprawdz_podwojne_ocenianie(
	  'gim_podw_kod.csv', 'gim_podw_kod_c.csv', dane, c('GH-P1A-14', 'GH-P1B-14'), '^pol.*_ost$'
	)
	sprawdz_podwojne_ocenianie(
	  'gim_podw_kod.csv', 'gim_podw_kod_c.csv', dane, c('GM-M1A-14', 'GM-M1B-14'), '^mat.*_ost$'
	)
}
}

