% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sprawdz_realizacje}
\alias{sprawdz_realizacje}
\title{Sprawdza poprawność pliku podsumowującego realizację badania}
\usage{
sprawdz_realizacje(plik, codebook, dane, przedmGim = NULL)
}
\arguments{
\item{plik}{ścieżka do pliku z danymi z realizacji (plik powinien mieć "format IBE")}

\item{codebook}{ścieżka do pliku codebooka danych z realizacji (plik powinien mieć "format IBE")}

\item{dane}{ramka danych z wynikami uczniów (wystarczą kolumny "id_szk", "id_kl", "wykl")}

\item{przedmGim}{opcjonalny wzór nazwy kolumny przechowującej w pliku realizacji liczbę zrealizowanych testów}
}
\value{
[logical] TRUE, gdy dane zweryfikowane poprawnie, FALSE w p.p.
}
\description{
Dopasowanie pomiędzy plikiem realizacji a plikiem wyników uczniów odbywa się
na podstawie zmiennych "id_szk" oraz "id_kl". Funkcja zakłada więc, że
zmienne takie istnieją w obydwu zbiorach.

Sprawdzane są:
\itemize{
	\item zawieranie wszystkich kombinacji {"id_szk", "id_kl"} w obydwu zbiorach
	\item zgodność liczby wystąpień danej kombinacji {"id_szk", "id_kl"} z wartością
  kolumny "l_bad" ("l_przedmGim" gdy podano argument "przedmGim")
	\item zależności pomiędzy kolumnami "l_uczkl", "l_wykl", "l_bad" ("l_przedmGim"
  gdy podano argument "przedmGim"), "l_ucz_szk", "l_dys", "l_dys_szk",
  "l_oddz"
}
}
\examples{
\dontrun{
	dane = wczytaj_ibe('spr.csv', 'spr_c.csv')
	sprawdz_realizacje('spr_realizacja.csv', 'spr_realizacja_c.csv', dane)
}
}

