% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stworz_skale}
\alias{stworz_skale}
\title{Tworzy w bazie nowa skale i zwraca jej id_skali}
\usage{
stworz_skale(nazwa, opis, rodzajEgzaminu, czescEgzaminu, rok,
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{nazwa}{nazwa skali (zwyczajowo w formacie "ZESPOL;EGZ;LATA", np. "paou;s;2002-2011" lub "ewd;gh_h;2012")}

\item{opis}{bardziej dokladny opis skali (opcjonalny, mozna ustawic na "")}

\item{rodzajEgzaminu}{jesli skala dotyczy jednej konkretnej czesci egzaminu, rodzaj tego egzaminu (patrz http://zpd.ibe.edu.pl/doku.php?id=czesci_egzaminu)}

\item{czescEgzaminu}{jesli skala dotyczy jednej konkretnej czesci egzaminu, czesc tego egzaminu (patrz http://zpd.ibe.edu.pl/doku.php?id=czesci_egzaminu)}

\item{rok}{jesli skala dotyczy jednej konkretnej czesci egzaminu, rok tego egzaminu (patrz http://zpd.ibe.edu.pl/doku.php?id=czesci_egzaminu)}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC, ktorego nalezy uzyc}
}
\value{
[numeric] id_skali utworzonej skali
}
\description{
_
}

