% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stworz_test}
\alias{stworz_test}
\title{Tworzy nowy test (także jako kompilację wskazanych testów)}
\usage{
stworz_test(ewd, opis, data, idTestow = NULL, zrodloDanychODBC = "EWD")
}
\arguments{
\item{ewd}{czy będzie to test ewd [logical]}

\item{opis}{opis testu}

\item{data}{data testu}

\item{idTestow}{wektor id_testu testów, których kryteria oceny mają zostać skopiowane do tworzonego testu}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC, ktorego nalezy uzyc}
}
\value{
[numeric] id_testu utworzonego testu
}
\description{
_
}

