% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wczytaj_ibe}
\alias{wczytaj_ibe}
\title{Funkcja wczytująca wskazany zbiór danych w formacie zgodnym ze standardami IBE}
\usage{
wczytaj_ibe(plik_danych, ksiazka_kodow, oo = NA, tp = NA, nd = NA,
  kolumna_skala = 6, kolumna_etykiety = 11, kolumna_typ = 5,
  kolumna_rozmiar = 7, kolumna_dokladnosc = 8, postep = F)
}
\arguments{
\item{plik_danych}{nazwa pliku ze zbiorem danych w formacie zgodnym ze standardami IBE do wczytania}

\item{ksiazka_kodow}{nazwa pliku z książką kodów w formacie zgodnym ze standardami IBE opisującej wczytywany zbiór danych}

\item{oo}{etykieta, jaka zostanie przypisana brakom danych IBE 'odmowa odpowiedzi/nie zaznaczono żadnej odpowiedzi' lub NA, jeśli danybrak danych IBE ma zostać skonwertowany na brak danych R (patrz opis)}

\item{tp}{etykieta, jaka zostanie przypisana brakom danych IBE 'trudno powiedzieć/wybrano więcej niż jedną odpowiedź' lub NA, jeśli dany brak danych IBE ma zostać skonwertowany na brak danych R (patrz opis)}

\item{nd}{etykieta, jaka zostanie przypisana brakom danych IBE 'nie dotyczy' lub NA, jeśli dany brak danych IBE ma zostać skonwertowany na brak danych R (patrz opis)}

\item{kolumna_skala}{numer kolumny książki kodów zawierającej opis skali zmiennych (jeśli książka kodów jest ściśle zgodna ze standardami IBE, to będzie to kolumna 6)}

\item{kolumna_etykiety}{numer kolumny książki kodów zawierającej etykiety zmiennych (jeśli książa kodów jest ściśle zgodna ze standardami IBE, to będzie to kolumna 11)}

\item{kolumna_typ}{numer kolumny książki kodów zawierającej typy zmiennych (jeśli książka kodów jest ściśle zgodna ze standardami IBE, to będzie to kolumna 5)}

\item{kolumna_rozmiar}{numer kolumny książki kodów zawierającej rozmiar zmiennych (jeśli książka kodów jest ściśle zgodna ze standardami IBE, to będzie to kolumna 7)}

\item{kolumna_dokladnosc}{numer kolumny książki kodów zawierającej dokładność zmiennych będących liczbami rzeczywistymi (jeśli książka kodów jest ściśle zgodna ze standardami IBE, to będzie to kolumna 8)}

\item{postep}{czy wyświetlać informacje o liczbie przetworzonych zmiennych}
}
\value{
[data.frame] wczytane dane
}
\description{
Funkcja wczytująca wskazany zbiór danych w formacie zgodnym ze standardami IBE.
Poza wczytaniem samych danych podczytuje z książki kodowej danego zbioru danych
etykiety zmiennych oraz skale.
}
\details{
Ponieważ R rozróżnia tylko skale nominalną, porządkową i całą resztę i tylko
dwa pierwsze rodzaje mogą być etykietowane, to mapowanie przebiega następująco:
\itemize{
	\item zmienne posiadające listę etykiet i mające wg książki kodowej skalę nominalną
  lub dychotomiczną konwertowane są na zmienne nominalne R-a (oraz przypisywane
  są im etykiety)
	\item zmienne posiadające listę etykiet i mające wg ksiązki kodowej skalę inną niż
  nominalna lub dychotomiczna konwertowane są na zmienne porządkowe R-a (oraz
  przypisywane są im etykiety) bez zwracania uwagi na to, jaką konkretnie skalę
  miały określoną w książce kodowej (!!!)
	\item zmienne nie posiadające listy etykiet, a mające wg książki kodowej skalę
  nominalną lub dychotomiczną konwertowane są na zmienne nominalne R-a
  z etykietami równymi liście ich unikalnych wartości
	\item zmienne nie posiadające listy etykiet, a mające wg książki kodowej skalę
  porządkową konwertowane są na zmienne porządkowe R-a z etykietami równymi
  liście ich unikalnych wartości
	\item pozostałe zmienne nie posiadające w książce kodowej listy etykiet nie są
  w żaden sposób modyfikowane
}

Druga kwestia, którą trzeba uwzględnić w R to to, że obsługuje on tylko jeden
brak danych (NA), podczas gdy standardy IBE rozróżniają trzy typy braków danych
(nie dotyczy, trudno powiedzieć/zaznaczono więcej niż jedną odpowiedź, odmowa
odpowiedzi/nie wybrano żadnej odpowiedzi). Dla każdej wartości specjalnej IBE
możliwe są dwa scenariusze:
\itemize{
	\item mapowanie wartości specjalnej IBE na NA;
	\item pozostawienie wartości specjalnych jako zwykłej wartości, przy czym jeśli
  zmienna ma określone etykietowanie, to pierwotna wartość kodu nie zostanie
  zachowana (etykietować w R można tylko zmienne nominalne i porządkowe, a tego
  rodzaju zmienne R przekodowuje na kolejne wartości od 1 począwszy i przypisuje
  każdej wartości etykietę pochodzącą z pierwotnej wartości zmiennej) i z góry
  znana będzie tylko wartość etykiety
}
Z tego powodu dla każdej z trzech wartości specjalnych IBE określających powód
braku danych dostępny jest przełącznik umożliwiający wskazanie etykiety, jaka
zostanie przypisana danemu brakowi danych lub wartości NA, jeśli ma on zostać
skonwertowany na brak danych R
}
\examples{
\dontrun{
	wczytaj_ibe('zr_spr.csv', 'zr_spr_c.csv', oo=-1, tp=-2)
}
}

