% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zapisz_spss}
\alias{zapisz_spss}
\title{Zapisuje dane jako plik SAV korzystając z PSPP}
\usage{
zapisz_spss(dane, plik, katalogTmp = NULL, sprzatnij = TRUE)
}
\arguments{
\item{dane}{ramka danych do zapisania}

\item{plik}{nazwa docelowego pliku SAV}

\item{katalogTmp}{katalog, w którym zapisane zostaną tymczasowy plik CSV i skrypt SPS}

\item{sprzatnij}{czy usuwać pliki tymczasowe; jeśli nie zostaną zwrócone ścieżki do nich}
}
\value{
[vector] NULL lub ścieżki do plików tymczasowych (patrz parametr "sprzatnij")
}
\description{
Zapis realizowany jest następująco:
\itemize{
 \item dane eksportowane są do pliku CSV;
	\item wtorzony jest skrypt SPSS-a, który wczyta dane z pliku CSV i zapisze
       je do pliku SAV;
	\item PSPP wywoływany jest wsadowo z przygotowanym skryptem;
	\item plik CSV jest usuwany.
}
W pliku SAV stosowane są tylko trzy typy zmiennych:
\itemize{
 \item string N, gdzie N jest odczytywane za pomocą stringr::stringi::stri_length()
 \item numeric N,M, gdzie N i M są odczytywane w przemyślny sposób
}
}
\examples{
\dontrun{
	zapisz_spss(mtcars, 'mtcars.sav')
}
}

