% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zastosuj_skale.R
\name{zastosuj_skale}
\alias{zastosuj_skale}
\title{Pobiera wyniki testow (odpowiedzi) w postaci dlugiej}
\usage{
zastosuj_skale(dane, src, idSkali, skroc = TRUE)
}
\arguments{
\item{dane}{wynik dzialania dowolnej z funkcji pobierz_wyniki_...()}

\item{src}{uchwyt źródła danych dplyr-a}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy stosować skróty skal}
}
\description{
Poniewaz skale definiuja przeksztalcenia wynikow w punktach (i nie da sie 
tych przeksztalcen zastosowac dla dystraktorow), zastosowanie skali powoduje
ze zwrocony zbior danych pozbawiony bedzie kolumny "odpowiedz"

Zastosowanie skali spowoduje rowniez automatyczne przyciecie kryteriow 
jedynie do tych, ktore naleza do wskazanej skali.
}

