% Generated by roxygen2 (4.0.2): do not edit by hand
\name{znajdz_testy}
\alias{znajdz_testy}
\title{Wyszukuje w bazie testy spełniające podane kryteria}
\usage{
znajdz_testy(opis = NULL, rodzajEgzaminu = NULL, czescEgzaminu = NULL,
  rok = NULL, arkusz = NULL, prefiks = NULL, data = NULL, ewd = NULL,
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{opis}{opis testu (w wypadku testów egzaminacyjnych złączany z rodzaju, części egzaminu i roku)}

\item{rodzajEgzaminu}{rodzaj egzaminu}

\item{czescEgzaminu}{część egzaminu}

\item{rok}{rok testu lub egzaminu}

\item{arkusz}{arkusz egzaminacyjny}

\item{prefiks}{prefiks części egzaminu}

\item{data}{data testu lub egzaminu}

\item{ewd}{czy test EWD}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC, ktorego nalezy uzyc}
}
\value{
[data.frame] ramka danych z wyszukanymi testami
}
\description{
Wyszukuje testy na podstawie dowolnej kombinacji cech przekazanych jako argumenty.

Każdy argument może być wyrażeniem regularnym.
}

