% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsumuj_punkty.R
\name{zsumuj_punkty}
\alias{zsumuj_punkty}
\title{Sumuje punkty w ramach testu}
\usage{
zsumuj_punkty(dane, usunKryteria = TRUE)
}
\arguments{
\item{dane}{wynik dzialania dowolnej z funkcji pobierz_wyniki_...()}

\item{usunKryteria}{czy usuwac ze zbioru kolumny z wynikami za poszczegolne kryteria}
}
\description{
Parametr "usunKryteria" dziala tylko dla danych w postaci szerokiej. Dla 
danych w postaci dlugiej nastepuje agregacja, ktora sila rzeczy usuwa
informacje o poszczegolnych kryteriach.
}
