% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agreguj_pwe.R
\name{agreguj_pwe}
\alias{agreguj_pwe}
\title{Oblicza PWE na podstawie PV agregujac wg wskazanych zmiennych}
\usage{
agreguj_pwe(dane, kolGrupy, wariancjaPop = 0, bladZrwn = 0,
  kolPV = "^[pP][vV]_[0-9]+|wynik$", kolNrPV = "nr_pv")
}
\arguments{
\item{dane}{macierz z wartosciami PV do zagregowania}

\item{kolGrupy}{nazwa kolumny (lub wektor nazw kolumn), wg których PWE mają zostać zagregowane}

\item{wariancjaPop}{wariancja wyników w populacji - patrz opis funkcji}

\item{bladZrwn}{błąd zrównywania - patrz opis funkcji}

\item{kolPV}{wyrażenie regularne dopasowujące kolumny z wartościami PV}

\item{kolNrPV}{nazwa kolumny z numerem PV (gdy dane w postaci długiej)}
}
\value{
[data.frame] zagregowane wartości PWE
}
\description{
_
}

