% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_dane_uczniowie_testy}
\alias{get_pupils_tests_data}
\alias{pobierz_dane_uczniowie_testy}
\title{Pobiera informacje o uczniach specyficzne dla testu / egzaminu}
\usage{
pobierz_dane_uczniowie_testy(src, daneOsobowe = FALSE)

get_pupils_tests_data(src, daneOsobowe = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{daneOsobowe}{czy pobierac z bazy takze dane osobowe (wymaga specjalnych uprawnien)}
}
\description{
Pobiera informacje o uczniach specyficzne dla testu / egzaminu
}

