% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pobierz_dane_uczniowie_testy.R
\name{pobierz_dane_uczniowie_testy}
\alias{pobierz_dane_uczniowie_testy}
\title{Pobiera informacje o uczniach specyficzne dla testu / egzaminu}
\usage{
pobierz_dane_uczniowie_testy(src, daneOsobowe = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{daneOsobowe}{czy pobierac z bazy takze dane osobowe (wymaga specjalnych uprawnien)}
}
\description{
Pobiera informacje o uczniach specyficzne dla testu / egzaminu
}

