% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_kryteria_oceny}
\alias{pobierz_kryteria_oceny}
\title{Pobiera informacje o kryteriach oceny i pytaniach}
\usage{
pobierz_kryteria_oceny(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera informacje o kryteriach oceny i pytaniach
}

