% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_oszacowania_uczniow.R
\name{pobierz_oszacowania_uczniow}
\alias{pobierz_oszacowania_uczniow}
\title{Pobiera oszacowania umiejętności uczniów}
\usage{
pobierz_oszacowania_uczniow(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera oszacowania umiejętności uczniów
}

