% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobierz_parametry.R
\name{pobierz_parametry}
\alias{pobierz_parametry}
\title{Pobiera parametry kryteriów i pseudokryteriów obliczone w skalowaniach}
\usage{
pobierz_parametry(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera parametry kryteriów i pseudokryteriów obliczone w skalowaniach
}

