% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_schematy_pkt.R
\name{pobierz_schematy_pkt}
\alias{pobierz_schematy_pkt}
\title{Pobiera dozwolone wartości punktowe kryteriów oceny}
\usage{
pobierz_schematy_pkt(src)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}
}
\description{
Pobiera dozwolone wartości punktowe kryteriów oceny
}
