% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_skale}
\alias{pobierz_skale}
\title{Pobiera skale i skalowania}
\usage{
pobierz_skale(src, doPrezentacji = TRUE, czyKtt = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{doPrezentacji}{czy wyświetlać tylko skale i skalowania oznaczone do
publicznej prezentacji}

\item{czyKtt}{czy czy wyświetlać skale i skalowania KTT}
}
\description{
Pobiera listę skal połączoną z listą skalowań oraz testami powiązanymi z daną
skalą. Tak więc każda skala występować będzie tyle razy, ile wynosi iloczyn
liczby testów, z którymi jest powiązana oraz liczby skalowań, które na niej
bazują.

Domyślnie pomijane są skale i skalowania, które nie zostały oznaczone jako
do_prezentacji. W większości wypadków powoduje to, że danej skali przypisane
jest co najwyżej jedno skalowanie, co powinno ułatwiać wybór poszukiwanego
skalowania. Jeśli chcemy uzyskać listę wszystkich skal, należy parametr
"doPrezentacji" ustawić na NA.

Domyślnie pomijane są wszystie skale i skalowania KTT. Aby wyszukać skale i
skalowania KTT, ustaw parametr "ktt" na TRUE.
}

