% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_uczniow}
\alias{get_pupils}
\alias{pobierz_uczniow}
\title{Pobiera informacje o uczniach}
\usage{
pobierz_uczniow(src, daneOsobowe = FALSE)

get_pupils(src, daneOsobowe = FALSE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{daneOsobowe}{czy pobierac z bazy takze dane osobowe (wymaga specjalnych uprawnien)}
}
\description{
Pobiera informacje o uczniach
}

