% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_wyniki_zrownywania.R
\name{pobierz_wyniki_zrownywania}
\alias{pobierz_wyniki_zrownywania}
\title{Pobiera ramke danych z wynikami egzaminacyjnymi testow zrownujacych}
\usage{
pobierz_wyniki_zrownywania(src, rodzajEgzaminu, rok, punktuj = TRUE,
  idSkali = NULL, skroc = TRUE)
}
\arguments{
\item{src}{uchwyt źródła danych dplyr-a}

\item{rodzajEgzaminu}{rodzaj egzaminu, ktorego wyniki maja zostac pobrane}

\item{rok}{rok, z ktorego dane maja zostac pobrane}

\item{punktuj}{wybor, czy dane maja byc pobrane w postaci dystraktorow, czy punktow}

\item{idSkali}{identyfikator skali, ktora ma zostac zastosowana do danych}

\item{skroc}{czy do danych zastosowac skrocenia skal opisane w skali}
}

