% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polacz.R
\name{polacz}
\alias{polacz}
\title{Zwraca połączenie dplyr-a z bazą danych ZPD}
\usage{
polacz(dbname = "ewd", host = "ewd.ibe.edu.pl", port = 5432,
  user = "ewd_baza", password = "CalEBo9")
}
\arguments{
\item{dbname}{nazwa bazy danych}

\item{host}{adres serwera bazy danych}

\item{port}{port, na którym nasłuchuje serwer baz danych}

\item{user}{użytkownik}

\item{password}{hasło}
}

