% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_pupils_tests_data}
\alias{get_exam_results}
\alias{get_pupils}
\alias{get_pupils_estimates}
\alias{get_pupils_tests_data}
\alias{get_schools}
\alias{get_test_results}
\alias{get_tests}
\alias{polacz}
\alias{zpd_connect}
\title{Zwraca połączenie dplyr-a z bazą danych ZPD}
\usage{
get_pupils_tests_data(src)

get_pupils_estimates(src)

get_schools(src)

get_tests(src)

get_pupils(src)

get_exam_results(src, rodzajEgzaminu, czescEgzaminu, rokEgzaminu, czyEwd,
  punktuj = TRUE, idSkali = NULL, skroc = FALSE)

get_test_results(src, idTestu, punktuj = TRUE, idSkali = NULL,
  skroc = FALSE)

polacz(dbname = "ewd", host = "ewd.ibe.edu.pl", port = 5432,
  user = "ewd_baza", password = "CalEBo9")

zpd_connect(dbname = "ewd", host = "ewd.ibe.edu.pl", port = 5432,
  user = "ewd_baza", password = "CalEBo9")
}
\arguments{
\item{dbname}{nazwa bazy danych}

\item{host}{adres serwera bazy danych}

\item{port}{port, na którym nasłuchuje serwer baz danych}

\item{user}{użytkownik}

\item{password}{hasło}
}
\description{
Zwraca połączenie dplyr-a z bazą danych ZPD
}

