% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polacz.R
\name{polacz}
\alias{polacz}
\title{Zwraca połączenie dplyr-a z bazą danych ZPD}
\usage{
polacz(
  dbname = "ewd",
  host = "127.0.0.1",
  port = 5432,
  user = "mzoltak",
  password = "oT_us5f29YRQ"
)
}
\arguments{
\item{dbname}{nazwa bazy danych}

\item{host}{adres serwera bazy danych}

\item{port}{port, na którym nasłuchuje serwer baz danych}

\item{user}{użytkownik}

\item{password}{hasło}
}
\description{
Zwraca połączenie dplyr-a z bazą danych ZPD
}
