% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sprawdz_skroty_skal}
\alias{sprawdz_skroty_skal}
\title{Sprawdza poprawnosc i tworzy w bazie brakujace skroty skal}
\usage{
sprawdz_skroty_skal(skroty, P)
}
\arguments{
\item{skroty}{wektor wartosci id_skrotu}

\item{P}{obiekt otwartego polaczenie ODBC}
}
\value{
[logical] TRUE w wypadku powodzenia funkcji
}
\description{
_
}

