% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sprawdz_zrownanie}
\alias{sprawdz_zrownanie}
\title{Weryfikuje, czy dane zrownania zostaly poprawnie zaimportowane do bazy}
\usage{
sprawdz_zrownanie(rodzajEgzaminu, rok, kolumnaId, zrodloDanychODBC = "EWD")
}
\arguments{
\item{rodzajEgzaminu}{rodzaj egzaminu}

\item{rok}{rok badania zrownujacego}

\item{kolumnaId}{nazwa kolumny przechowujacej "zewnetrzne id"}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC}
}
\value{
[logical] TRUE, gdy dane zweryfikowane poprawnie, FALSE w p.p.
}
\description{
Funkcja:
\itemize{
	\item pobiera dane wskazanego zrownania (rodzaj egzaminu i rok) z bazy
	\item dla kazdego testu zrownujacego sprawdza jego dane w bazie z plikiem "id_testu.csv"
  pliki te powinny sie znajdowac w katalogu roboczym
}
}
\examples{
\dontrun{
	sprawdz_zrownanie('sprawdzian', 2014, 'zbiór.zrównywanie')
}
}

