% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stworz_skale.R
\name{stworz_skale}
\alias{stworz_skale}
\title{Tworzy w bazie nowa skale i zwraca jej id_skali}
\usage{
stworz_skale(nazwa, opis, rodzaj, doPrezentacji, idTestow,
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{nazwa}{nazwa skali (zwyczajowo w formacie "ZESPOL;EGZ;LATA", np. "paou;s;2002-2011" lub "ewd;gh_h;2012")}

\item{opis}{bardziej dokladny opis skali (opcjonalny, mozna ustawic na "")}

\item{rodzaj}{rodzaj skali (ewd/zrównywanie/ktt)}

\item{doPrezentacji}{czy skala ma być oznaczona jako przeznaczona do prezentacji}

\item{idTestow}{wektor id testów, z którymi ma być powiązana skala}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC, ktorego nalezy uzyc}
}
\value{
[numeric] id_skali utworzonej skali
}
\description{
_
}

