% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wczytaj_spss}
\alias{wczytaj_spss}
\title{Wczytuje wskazany plik SPSS stosujac standardowe przeksztalcenia}
\usage{
wczytaj_spss(plik)
}
\arguments{
\item{plik}{sciezka do pliku SPSS}
}
\value{
[data.frame] wczytane dane
}
\description{
Funkcja:
\itemize{
	\item wczytuje plik SPSS z to.data.frame=T, use.missings=F, trim.factor.names=T
	\item zamienia nazwy wszystkich zmiennych na pisane malymi literami
	\item zamienia wszystkie factor-y na zwykle zmienne typu character
	\item wykonuje str_trim() na wszystkich zmiennych
}
}
\examples{
\dontrun{
	wczytaj_spss('gim10_07.sav')
}
}

