% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wydziel_testy}
\alias{wydziel_testy}
\title{Dzieli plik danych z badan zrownujacych na testy}
\usage{
wydziel_testy(dane, zeszyt, zeszytBaza, kolIdZeszytu, kolUdzial, kolIdSzkOKE,
  prefiksZadan, rok, zrodloDanychODBC = "EWD")
}
\arguments{
\item{dane}{ramka danych}

\item{zeszyt}{string z nazwa zeszytu (taka, jak w kolumnie wskazywanej przez "kolIdZeszytu")}

\item{zeszytBaza}{string z nazwa zeszytu w bazie danych ("zrownywanie;RODZAJ_EGZ;ROK;KOD")}

\item{kolIdZeszytu}{numer kolumny w "dane" przechowujacej nazwy zeszytow}

\item{kolUdzial}{numer kolumny w "dane" przechowujacej informacje o udziale ucznia w badaniu}

\item{kolIdSzkOKE}{numer kolumny w "dane" przechowujace id OKE szkoly}

\item{prefiksZadan}{prefiks kolumn zawierajacych wyniki zadan dla danego zeszytu}

\item{rok}{rok testu zrownujacego (niezbedny do skonstruowania identyfikatora 'zrwn')}

\item{zrodloDanychODBC}{nazwa zrodla danych ODBC}
}
\value{
[void]
}
\description{
Funkcja:
\itemize{
	\item filtruje rekordy pasujace do zadanego zeszytu
	\item zmienia nazwy kolumn z wynikami zadan na id_kryteriow
	\item dodaje kolumne "zrwn" zawierajace kolumne "id_ucz" przeksztalcona do postaci uzywanego w bazie identyfikatora "zrwn"
 \item dubluje kolumne "id_szk" pod nazwa "id_szk_oke"
 \item zapisuje tak odfiltrowany i skonwertowany zbior pod nazwa "id_testu.csv"
}
}
\examples{
\dontrun{
	# podział wyników zrównywania sprawdzianu 2014 na testy
	dane = wczytaj_ibe('spr.csv', 'spr_c.csv', oo='-2', tp='-1')
	names(dane) = tolower(names(dane))
	dane$"zbiór zrównywanie" = paste('s 2014', dane$id_ucz, sep='|')
	dane$klasa = dane$id_kl
	dane$kod_u = sub('^.*-[^0-9]', '\\\\1', dane$id_ucz)
	kolUdzial = grep('^udzial$', names(dane))
	for(i in 1:1){
		for(j in c('A', 'B')){
			wydziel_testy(
					dane = dane,
					zeszyt = sprintf('S-\%d\%s-14', i, j),
					zeszytBaza = sprintf('zrównywanie;sprawdzian;2014;\%d\%s', i,  j),
					kolIdZeszytu = grep('^id_zt$', names(dane)),
					kolUdzial = kolUdzial,
					kolIdSzkOKE = grep('^id_szk$', names(dane)),
					prefiksZadan = sprintf('z\%d_[1-9]', i),
					rok = 2014
			)
		}
	}
}
}

