% Generated by roxygen2 (4.0.1): do not edit by hand
\name{znajdz_pseudokryteria}
\alias{znajdz_pseudokryteria}
\title{Odnajduje i tworzy pseudokryteria na podstawie list ich elementow}
\usage{
znajdz_pseudokryteria(kryteria, opisy, P)
}
\arguments{
\item{kryteria}{macierz id_kryterium - kazdy wiersz opisuje jedno pseudokryterium}

\item{opisy}{wektor opisow pseudokryteriow}

\item{P}{obiekt otwartego polaczenie ODBC}
}
\value{
[numeric vector] wektor odnalezionych i/lub utworzonych id_pseudokryteriow
}
\description{
_
}

